requirements.txt
# Core runtime dependencies
deepgram-sdk
google-genai
numpy
sounddevice
